requirements.txt
Flask==3.0.0
flask_sqlalchemy==3.1.1
lxml==4.9.3
PyJWT==2.8.0
Werkzeug==3.0.1